+++ index.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Awesome Books</title>
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="index.css">
 </head>
 <body>
   <header>
@@ -49,7 +49,7 @@
          <p>Copyright</p>
       </div>
     </footer>
-    <script type="module" src="script.js"></script>
-    <script type="module" src="date_app.js"></script>
+    <script type="module" src="index.js"></script>
+    <!-- <script type="module" src="./modules/date.js"></script> -->
 </body>
 </html>
