+++ modules/bookClass.js
@@ -1,0 +1,7 @@
+export default class Book {
+  constructor(title, author) {
+    this.title = title;
+    this.author = author;
+    this.id = Date.now();
+  }
+}
