--- style.css
@@ -1,172 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-body {
-  font-family: 'Trebuchet MS', 'Lucida Sans Unicode', 'Lucida Grande', 'Lucida Sans', Arial, sans-serif;
-  width: 100%;
-}
-
-.books-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 10px;
-  padding: 0 10%;
-}
-
-.List {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  border: 2px solid black;
-  padding: 10px;
-}
-
-.form {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 20px;
-  width: 50%;
-}
-
-.single-book {
-  height: 35px;
-  display: flex;
-  justify-content: space-between;
-  padding: 5px;
-}
-
-.single-book:nth-child(odd) {
-  background: rgb(189, 183, 183);
-}
-
-.bio {
-  display: flex;
-  gap: 5px;
-  align-items: center;
-}
-
-.add-books {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  padding: 5px;
-  margin-top: 200px;
-}
-
-.form input {
-  border: 3px solid black;
-  height: 40px;
-  width: 100%;
-  padding-left: 10px;
-  font-weight: 900;
-}
-
-.form button {
-  display: flex;
-  align-self: flex-end;
-}
-
-.title-form,
-.title-contact {
-  margin-bottom: 50px;
-  font-size: 2rem;
-}
-
-#add {
-  width: 50px;
-  border: 2px solid #000;
-  font-weight: 900;
-  background-color: #fff;
-  border-right: 4px solid #000;
-  border-bottom: 4px solid #000;
-}
-
-#add,
-.rmvBtn {
-  border: 2px solid #000;
-  font-weight: 900;
-  background-color: #fff;
-  border-right: 4px solid #000;
-  border-bottom: 4px solid #000;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 3px;
-}
-
-.title {
-  font-weight: 700;
-}
-
-.uppercase {
-  text-transform: capitalize;
-}
-
-#Cr {
-  border: 2px solid #000;
-  font-weight: 700;
-  position: absolute;
-  bottom: 0;
-  margin: 0 10px;
-  margin-bottom: 10px;
-  width: 98%;
-  padding: 10px 0;
-  padding-left: 10px;
-}
-
-header {
-  display: flex;
-  justify-content: space-between;
-  border: 2px solid #000;
-  margin: 10px;
-  padding: 10px 30px;
-}
-
-#nav-list {
-  list-style-type: none;
-  display: flex;
-}
-
-#nav-title {
-  font-size: 1.3rem;
-}
-
-nav > ul > li > a {
-  text-decoration: none;
-  color: #000;
-  font-weight: 700;
-  font-size: 1.3rem;
-  border-right: 2px solid black;
-  padding: 0 10px;
-}
-
-ul > li:last-child > a {
-  border-right: none;
-}
-
-#date {
-  display: flex;
-  justify-content: flex-end;
-  margin-right: 20px;
-  font-weight: 700;
-}
-
-.contSection {
-  display: none;
-  flex-direction: column;
-  font-weight: 900;
-  align-items: center;
-  margin-top: 100px;
-  font-size: larger;
-}
-
-#cont-list {
-  margin: 30px;
-}
